Okay, this is the real log entry: Author: costar Date: 2000-06-30 07:31:18 GMT Okay, this is the real log entry:
Added kernel-source-2.2.17, so that people with 1 CD (from a Linux show or
similar) can at least compile their own kernel.

Saw that twm was the only window manager on CD 1, which isn't quite
sufficient for most people. So added enlightenment and wmaker, which seem
to be the most popular (and work great with Gnome). Added twm as well,
because otherwise apt will think the window-manager is already there and
leaves twm at the end, which is bad because some startx-like scripts and
printed manuals/books assume it's present. Also added fvwm and fvwm95 for
those running a 386 w/8MB (on which E/wmaker take >5 mins to start up).