CONF.sh: Author: philh Date: 2000-08-25 16:54:03 GMT CONF.sh: VERBOSE_MAKE variable added.  allows the verbosity of the make process to be controlled, by determining if the lines start with a @ or not ATTEMPT_FALLBACK variable added.  allows one to control whether build_all.sh should bother to attempt to build a simple CD if the official build fails.
Makefile:
o  added "set -e;" to the start of most of the blocks of shell commands to
ensure that if one of the scripts fails, that fact is not ignored
o  replaced many of the @'s with $(Q) and ensured that Q=@ by default.
This allows one to see what make is actually doing when it all goes
horribly wrong -- VERBOSE_MAKE in CONF.sh is used to control this

build_all.sh
o  made it stop on error dependant on ATTEMPT_FALLBACK from CONF.sh

tools/pi-makelist:
These are the changes by J.A. Bezemer to get rid of files from the list
that are not available on ftp mirrors

tools/boot/potato/boot-alpha
check to see if the boot1 directory already exists, and skip its creation
if so.  This is required to deal with the fact that this script gets called
for CD1 & CD1_NONUS, and making the directories twice fails, and is pointless