Change the way to insert the [Image] section on jigdos ... Author: manty Date: 2002-07-03 12:01:50 GMT
Change the way to insert the [Image] section on jigdos so that jigdo-file
generates the md5sum of the template and adds it to the [Image] section.
A side efect is that the generation date apearing on the Info changes from
being the end date to being the start date. If this matters tell me and
I'll change it.